# libs/web/api-client/src/lib/api/models/paginated-documented.ts
/* tslint:disable */
/* eslint-disable */
import { PaginatedMetaDocumented } from '../models/paginated-meta-documented';
export interface PaginatedDocumented {
  data: Array<{
}>;
  meta: PaginatedMetaDocumented;
}