# libs/web/api-client/src/lib/api/models/paginated-meta-documented.ts
/* tslint:disable */
/* eslint-disable */
export interface PaginatedMetaDocumented {
  currentPage: number;
  itemsPerPage: number;
  sortBy?: Array<Array<(string | 'ASC' | 'DESC')>>;
  totalItems: number;
  totalPages: number;
}