# libs/web/api-client/src/lib/swagger.json
{"openapi":"3.0.0","paths":{"/api/units":{"get":{"operationId":"UnitsController_getAll","summary":"get all units","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PaginatedDocumented"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Unit"}}}}]}}}}},"tags":["Units"]},"post":{"operationId":"UnitsController_create","summary":"create a new unit","parameters":[],"requestBody":{"required":true,"description":"Json structure for unit object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUnitDto"}}}},"responses":{"201":{"description":"Successful create","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unit"}}}}},"tags":["Units"]}},"/api/units/{id}":{"get":{"operationId":"UnitsController_getOne","summary":"get one unit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unit"}}}}},"tags":["Units"]},"delete":{"operationId":"UnitsController_delete","summary":"delete unit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Resource deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unit"}}}}},"tags":["Units"]}}},"info":{"title":"bakery-information-system-api","description":"[swagger file](http://localhost:3000/api-json)","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"PaginatedMetaDocumented":{"type":"object","properties":{"itemsPerPage":{"type":"number","title":"Number of items per page"},"totalItems":{"type":"number","title":"Total number of items"},"currentPage":{"type":"number","title":"Current requested page"},"totalPages":{"type":"number","title":"Total number of pages"},"sortBy":{"type":"array","title":"Sorting by columns","items":{"type":"array","items":{"oneOf":[{"type":"string"},{"type":"string","enum":["ASC","DESC"]}]}}}},"required":["itemsPerPage","totalItems","currentPage","totalPages"]},"PaginatedDocumented":{"type":"object","properties":{"data":{"title":"Array of entities","type":"array","items":{"type":"object"}},"meta":{"title":"Pagination Metadata","allOf":[{"$ref":"#/components/schemas/PaginatedMetaDocumented"}]}},"required":["data","meta"]},"Unit":{"type":"object","properties":{"code":{"type":"string"},"fullName":{"type":"string"},"id":{"type":"string"},"name":{"type":"string"}},"required":["code","fullName","id","name"]},"CreateUnitDto":{"type":"object","properties":{"code":{"type":"string"},"fullName":{"type":"string"},"name":{"type":"string"}},"required":["code","fullName","name"]}}}}